+++ src/main/kotlin/co/llective/hyena/api/api.kt
@@ -105,7 +105,7 @@ data class DataTriple(val columnId: Long, val columnType: BlockType, val data: O
 
 data class ScanResult(val data: List<DataTriple>)
 
-open class Column(val dataType: BlockType, val id: Int, val name: String) {
+open class Column(val dataType: BlockType, val id: Long, val name: String) {
     override fun toString(): String = "$name/$id ${dataType.name}"
 }
